backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
stripe==8.4.0
praw==7.8.1
openai==1.65.5
google-ads==26.0.0